+++ README.md
@@ -1,6 +1,9 @@
 # GuildedURLBot
 This is a guilded URL sniper, it snipes URLs not efficiently YET.
 
+# Libraries
+This uses Selinium, json, time, json, random
+
 # Running
 
 Create a json file called <code>Password.json</code><br>, make sure its a Capital Letter
@@ -10,3 +13,7 @@ Then run the json file and follow the instructions
 
 # Support
 If you would like to countribute to this project please do so, any help is wanted!
+
+# More Stuff
+
+I do not support andrew tate it is a joke.
